--- compute/cloud-client/src/main/java/compute/ipaddress/IpType.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package compute.ipaddress;
-
-public enum IpType {
-  INTERNAL("internal"),
-  EXTERNAL("external"),
-  IP_V6("ipv6");
-
-  private final String type;
-
-  IpType(String type) {
-    this.type = type;
-  }
-
-  public String getType() {
-    return type;
-  }
-}
